+++ src/app/(frontend)/(home)/_components/Events/EventCard.tsx
@@ -5,8 +5,8 @@ import Image from 'next/image';
 import { EventData } from '@/types/EventData';
 import { Button } from '@/components/ui/Button';
 import { motion, AnimatePresence } from 'framer-motion';
-import { Clock, DollarSign, MapPin } from 'lucide-react';
 import ArrowUp from '@/components/icons/ArrowUp';
+import EventInfo from "@/app/(frontend)/(home)/_components/Events/EventInfo";
 
 const days = ['Sunday', 'Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday', 'Saturday'];
 
@@ -21,12 +21,7 @@ export default function EventCard({ event, even, isPast }: EventCardProps) {
     const date = new Date(event.date);
     const disabled = event.locked || isPast;
 
-    const startTime = new Date(event.startTime);
-    const endTime = new Date(event.endTime);
-
-    function formatTime(date: Date) {
-        return date.toLocaleTimeString([], { hour: '2-digit', minute: '2-digit', hour12: false });
-    }
+    const dateString = `${date.getDate().toString().padStart(2, '0')}.${(date.getMonth() + 1).toString().padStart(2, '0')}.${date.getFullYear().toString().slice(-2)}`
 
     return (
         <div className="w-full">
@@ -42,9 +37,7 @@ export default function EventCard({ event, even, isPast }: EventCardProps) {
                         <div className="flex flex-col">
                             <h4>{days[date.getDay()]}</h4>
                             <h4>
-                                {date.getDate().toString().padStart(2, '0')}.
-                                {(date.getMonth() + 1).toString().padStart(2, '0')}.
-                                {date.getFullYear().toString().slice(-2)}
+                                {dateString}
                             </h4>
                             <Button
                                 variant="clear"
@@ -104,78 +97,65 @@ export default function EventCard({ event, even, isPast }: EventCardProps) {
                     >
                         <div className="grid grid-cols-1 md:grid-cols-[360px_1fr] gap-6 items-stretch">
                             {/* Tall poster */}
-                            <div className="relative w-full h-[520px] rounded-3xl overflow-hidden shadow-lg">
-                                <Image
-                                    src={event.image}
-                                    alt={event.imageAlt}
-                                    fill
-                                    priority
-                                    className={`object-cover ${event.locked ? 'blur-sm' : ''}`}
-                                    draggable={false}
-                                />
-                                {event.locked && (
+                            <h4 className="md:hidden flex font-bold">
+                                {!event.locked ? event.title : 'Locked Event'}
+                            </h4>
+                            <div className="flex justify-between gap-x-5">
+                                <div className="relative w-[185px] md:w-full h-[231px] md:h-[520px] rounded-3xl overflow-hidden shadow-lg">
                                     <Image
-                                        src="/images/home/lock.png"
-                                        alt="Locked"
-                                        width={80}
-                                        height={80}
+                                        src={event.image}
+                                        alt={event.imageAlt}
+                                        fill
+                                        priority
+                                        className={`object-cover ${event.locked ? 'blur-sm' : ''}`}
                                         draggable={false}
-                                        className="absolute inset-0 m-auto z-10"
                                     />
-                                )}
+                                    {event.locked && (
+                                        <Image
+                                            src="/images/home/lock.png"
+                                            alt="Locked"
+                                            width={80}
+                                            height={80}
+                                            draggable={false}
+                                            className="absolute inset-0 m-auto z-10"
+                                        />
+                                    )}
+
+                                </div>
+                                <EventInfo event={event} dateString={dateString} className="md:hidden" />
                             </div>
 
                             {/* Right panel matches image height */}
-                            <div className="flex flex-col w-full h-[520px] rounded-3xl !bg-transparent !border-none">
+                            <div className="flex flex-col w-full h-full md:h-[520px] rounded-3xl !bg-transparent !border-none">
                                 <div>
-                                    <div className="flex justify-between items-center">
-                                        <h4 className="text-xl md:text-2xl font-bold leading-tight">
-                                            {!event.locked ? event.title : 'Locked Event'}
-                                        </h4>
-                                        <h4>
-                                            {date.getDate().toString().padStart(2, '0')}.
-                                            {(date.getMonth() + 1).toString().padStart(2, '0')}.
-                                            {date.getFullYear().toString().slice(-2)}
-                                        </h4>
-                                    </div>
-                                    <div className="flex flex-col mt-5">
-                                        <div className="flex justify-between">
-                                            <span className="flex flex-row items-center">
-                                                <DollarSign /> {event.memberPrice} (Members)
-                                            </span>
-                                            <span className="flex flex-row items-center">
-                                                <Clock className="mr-1" />{' '}
-                                                {`${formatTime(startTime)} - ${formatTime(endTime)}`}
-                                            </span>
-                                        </div>
-                                        <div className="flex justify-between">
-                                            <span className="flex flex-row items-center">
-                                                <DollarSign /> {event.nonMemberPrice} (Non-Members)
-                                            </span>
-                                            <span className="flex flex-row items-center">
-                                                <MapPin className="mr-1" />
-                                                {event.location}
-                                            </span>
-                                        </div>
-                                    </div>
-                                    <div className="mt-4 text-gray-300 leading-relaxed">
-                                        {!event.locked
-                                            ? event.description
-                                            : 'This is a locked upcoming event, come back another time to find out what it is! 👀'}
-                                    </div>
+                                <div className="hidden md:flex justify-between items-center">
+                                    <h4 className="text-xl md:text-2xl font-bold leading-tight">
+                                        {!event.locked ? event.title : 'Locked Event'}
+                                    </h4>
+                                    <h4>
+                                        {dateString}
+                                    </h4>
                                 </div>
-
-                                {/* Button stuck bottom right */}
-                                <div className="mt-auto flex justify-end">
-                                    <Button
-                                        disabled={event.locked}
-                                        className="flex flex-row items-center gap-x-2"
-                                    >
-                                        Sign Up
-                                        <ArrowUp className="size-3" />
-                                    </Button>
+                                <EventInfo event={event} dateString={dateString} className="hidden md:flex" />
+                                <hr className="hidden md:flex mt-5" />
+                                <div className="mt-4 text-gray-300 leading-relaxed">
+                                    {!event.locked
+                                        ? event.description
+                                        : 'This is a locked upcoming event, come back another time to find out what it is! 👀'}
                                 </div>
                             </div>
+
+                            {/* Button stuck bottom right */}
+                            <div className="my-6 md:my-0 md:mt-auto flex justify-end">
+                                <Button
+                                    disabled={event.locked}
+                                    className="flex flex-row items-center gap-x-2"
+                                >
+                                    Sign Up
+                                    <ArrowUp className="size-3" />
+                                </Button>
+                            </div>
+                                </div>
                         </div>
                     </motion.div>
                 )}
